// src/domain/dtos/book.dto.ts
export interface BookDto {
  id: string;
  title: string;
  subtitle?: string;
  author: string;
  description?: string;
  publishedDate?: string;
  publisher?: string;
  isbn?: string;
  pageCount?: number;
  imageUrl?: string;
  googleBooksId?: string;
  userId: string;
}

export interface CreateBookDto {
  title: string;
  subtitle?: string;
  author: string;
  description?: string;
  publishedDate?: string;
  publisher?: string;
  isbn?: string;
  pageCount?: number;
  imageUrl?: string;
  googleBooksId?: string;
}

export interface UpdateBookDto extends Partial<CreateBookDto> {}